+++ tests/attributes/HasDataTest.php
@@ -12,7 +12,7 @@ use yii\ui\attributes\HasData;
 use yii\ui\exception\Message;
 use yii\ui\helpers\Attributes;
 use yii\ui\mixin\HasAttributes;
-use yii\ui\tests\providers\tag\attributes\DataProvider;
+use yii\ui\tests\providers\attributes\DataProvider;
 
 /**
  * Test suite for {@see HasData} trait functionality and behavior.
@@ -20,7 +20,7 @@ use yii\ui\tests\providers\tag\attributes\DataProvider;
  * Validates the management of the global HTML `data-*` attributes according to the HTML Living Standard specification.
  *
  * Ensures correct handling, immutability, and validation of `data-*` attributes in tag rendering, supporting both
- * `string` and `\Closure` values for dynamic data assignment.
+ * `string` and `Closure` values for dynamic data assignment.
  *
  * Test coverage.
  * - Accurate rendering of attributes with `data-*` attributes.